requirements.txt
mediapipe==0.8.9
numpy==1.20.0
progressbar33==2.4
opencv-python==4.5.4.58
numba==0.54.1